+++ Packs/ZeroFox/Classifiers/classifier-ZeroFox_Key_Incident_Incoming_Mapper.json
@@ -41,7 +41,7 @@
 			"internalMapping": {}
 		}
 	},
-	"name": "108e33e9-82ab-4ed5-8fe3-989830521c15",
+	"name": "ZeroFox Key Incidents (incoming)",
 	"type": "mapping-incoming",
 	"version": -1
 }
