+++ Packs/ZeroFox/IncidentFields/ZeroFoxKeyIncidents_Analysis.json
@@ -1,18 +1,17 @@
 {
-    "id": "incident_headline",
+    "id": "incident_analysis",
     "version": -1,
-    "fromVersion": "6.0.0",
-    "modified": "2025-04-01T14:36:27.595586447Z",
-    "content": true,
+    "modified": "2025-04-01T14:37:02.293030483Z",
     "itemVersion": "",
     "toServerVersion": "",
-    "name": "Headline",
+    "fromVersion": "6.0.0",
+    "name": "Analysis",
     "ownerOnly": true,
     "placeholder": "",
     "template": "",
     "description": "",
-    "cliName": "headline",
-    "type": "shortText",
+    "cliName": "analysis",
+    "type": "longText",
     "closeForm": false,
     "editForm": true,
     "required": false,
@@ -26,6 +25,7 @@
     "useAsKpi": false,
     "locked": false,
     "system": false,
+    "content": true,
     "group": 0,
     "hidden": false,
     "openEnded": false,
@@ -35,7 +35,7 @@
     "systemAssociatedTypes": null,
     "associatedToAll": false,
     "unmapped": false,
-    "unsearchable": false,
+    "unsearchable": true,
     "caseInsensitive": true,
     "columns": null,
     "defaultRows": null,
