+++ Packs/ZeroFox/IncidentFields/ZeroFoxKeyIncidents_Headline.json
@@ -1,17 +1,18 @@
 {
-    "id": "incident_analysis",
+    "id": "incident_headline",
     "version": -1,
-    "modified": "2025-04-01T14:37:02.293030483Z",
+    "fromVersion": "6.0.0",
+    "modified": "2025-04-01T14:36:27.595586447Z",
+    "content": true,
     "itemVersion": "",
     "toServerVersion": "",
-    "fromVersion": "6.0.0",
-    "name": "Analysis",
+    "name": "Headline",
     "ownerOnly": true,
     "placeholder": "",
     "template": "",
     "description": "",
-    "cliName": "analysis",
-    "type": "longText",
+    "cliName": "headline",
+    "type": "shortText",
     "closeForm": false,
     "editForm": true,
     "required": false,
@@ -25,7 +26,6 @@
     "useAsKpi": false,
     "locked": false,
     "system": false,
-    "content": true,
     "group": 0,
     "hidden": false,
     "openEnded": false,
@@ -35,7 +35,7 @@
     "systemAssociatedTypes": null,
     "associatedToAll": false,
     "unmapped": false,
-    "unsearchable": false,
+    "unsearchable": true,
     "caseInsensitive": true,
     "columns": null,
     "defaultRows": null,
