--- Packs/ZeroFox/ReleaseNotes/1_4_0.md
@@ -1,3 +1,0 @@
-#### Integrations
-##### ZeroFox_Key_Incidents
-- Creates application for ingesting ZeroFox Key Incidents as XSOAR Incident data
